+++ src/main/java/com/example/pigonair/domain/payment/entity/Payment.java
@@ -1,9 +1,9 @@
-package com.example.pigonair.payment.entity;
+package com.example.pigonair.domain.payment.entity;
 
-import com.example.pigonair.flight.entity.Flight;
-import com.example.pigonair.member.entity.Member;
-import com.example.pigonair.reservation.entity.Reservation;
-import com.example.pigonair.seat.entity.Seat;
+import com.example.pigonair.domain.member.entity.Member;
+import com.example.pigonair.domain.reservation.entity.Reservation;
+import com.example.pigonair.domain.flight.entity.Flight;
+import com.example.pigonair.domain.seat.entity.Seat;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
