+++ src/main/java/com/example/pigonair/domain/seat/repository/FlightRepository.java
@@ -1,8 +1,8 @@
-package com.example.pigonair.seat.repository;
+package com.example.pigonair.domain.seat.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.pigonair.flight.entity.Flight;
+import com.example.pigonair.domain.flight.entity.Flight;
 
 public interface FlightRepository extends JpaRepository<Flight,Long> {
 }
